modularize Python code.
